+++ optree/typing.py
@@ -36,6 +36,7 @@ from typing import (
     TypeVar,
     Union,
 )
+from typing_extensions import NamedTuple  # Generic NamedTuple: Python 3.11+
 from typing_extensions import OrderedDict  # Generic OrderedDict: Python 3.7.2+
 from typing_extensions import Self  # Python 3.11+
 from typing_extensions import TypeAlias  # Python 3.10+
@@ -43,13 +44,6 @@ from typing_extensions import Final, Protocol, runtime_checkable  # Python 3.8+
 
 from optree import _C
 from optree._C import PyTreeKind, PyTreeSpec
-
-
-try:
-    # pylint: disable-next=ungrouped-imports
-    from typing_extensions import NamedTuple  # Generic NamedTuple: Python 3.11+
-except ImportError:
-    from typing import NamedTuple  # type: ignore[assignment]
 
 
 __all__ = [
@@ -165,13 +159,13 @@ class PyTree(Generic[T]):  # pylint: disable=too-few-public-methods
         if len(item) != 2:
             raise TypeError(
                 f'{cls.__name__}[...] only supports a tuple of 2 items, '
-                f'a parameter and a string of type name. Got {item!r}.',
+                f'a parameter and a string of type name, got {item!r}.',
             )
         param, name = item
         if name is not None and not isinstance(name, str):
             raise TypeError(
                 f'{cls.__name__}[...] only supports a tuple of 2 items, '
-                f'a parameter and a string of type name. Got {item!r}.',
+                f'a parameter and a string of type name, got {item!r}.',
             )
 
         if (
@@ -244,7 +238,7 @@ class PyTreeTypeVar:
     def __new__(cls, name: str, param: type) -> TypeAlias:
         """Instantiate a PyTree type variable with the given name and parameter."""
         if not isinstance(name, str):
-            raise TypeError(f'{cls.__name__} only supports a string of type name. Got {name!r}.')
+            raise TypeError(f'{cls.__name__} only supports a string of type name, got {name!r}.')
         return PyTree[param, name]  # type: ignore[misc,valid-type]
 
     def __init_subclass__(cls, *args: Any, **kwargs: Any) -> NoReturn: