+++ tests/test_ops.py
@@ -1530,10 +1530,10 @@ def test_tree_flatten_one_level(tree, none_is_leaf, namespace):  # noqa: C901
     expected_leaves = optree.tree_leaves(tree, none_is_leaf=none_is_leaf, namespace=namespace)
     while stack:
         node = stack.pop()
-        counter = Counter()
+        counter = itertools.count()
         expected_children, one_level_treespec = optree.tree_flatten(
             node,
-            is_leaf=lambda x: counter.increment() > 1,  # noqa: B023
+            is_leaf=lambda x: next(counter) > 0,  # noqa: B023
             none_is_leaf=none_is_leaf,
             namespace=namespace,
         )